requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langfuse>=2.19.0
openai>=1.12.0
chromadb>=0.4.22
tiktoken>=0.5.2
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.29.0
pypdf>=4.0.1
faiss-cpu>=1.9.0

